feat: Inserir interfaces e estrutura de diretórios
