+++ crud-csharp-sqlserver/App.config
@@ -5,4 +5,11 @@
       connectionString="Data Source=localhost,1433;Initial Catalog=AppSingleton;User ID=sa;Password=Local@123456789;"
       providerName="System.Data.SqlClient" />
   </connectionStrings>
+  <!-- Parâmetros de buscas -->
+  <appSettings>
+    <add key="PageSize" value="10" />
+    <add key="PageNumber" value="1" />
+    <add key="OrderBy" value="Id" />
+    <add key="OrderDirection" value="ASC" />
+  </appSettings>
 </configuration>
