+++ crud-csharp-sqlserver/Data/IDbConnection.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Data.SqlClient;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace SingletonCRUD; 
+
+public interface IDbConnection : IDisposable
+{
+    SqlConnection GetConnection();          
+    public void Dispose();     
+    
+}
