+++ crud-csharp-sqlserver/Program.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using System.Security.Cryptography.X509Certificates;
+using System;
 using System.Data.SqlClient;
 
 namespace SingletonCRUD
@@ -7,14 +8,15 @@ namespace SingletonCRUD
     {
         static void Main(string[] args)
         {
-            DbOperatesDAO dbOperates = new DbOperatesDAO();
+
+            OperatesRepository repository = new OperatesRepository();
 
             Console.WriteLine("Inserindo dados...");
-            dbOperates.Insert("João", 25);
-            dbOperates.Insert("Maria", 30);
+            repository.Insert("Marcelo Augusto", 45);
+            repository.Insert("Angela Marques", 32);
 
             Console.WriteLine("\nExibindo todos os registros...");
-            SqlDataReader reader = dbOperates.SelectAll();
+            SqlDataReader reader = repository.SelectAll();
             while (reader.Read())
             {
                 Console.WriteLine($"Nome: {reader["Nome"]}, Idade: {reader["Idade"]}");
@@ -22,10 +24,10 @@ namespace SingletonCRUD
             reader.Close();
 
             Console.WriteLine("\nAtualizando dados...");
-            dbOperates.Update(1, "José", 28);
+            repository.Update(1, "José", 28);
 
             Console.WriteLine("\nExibindo registro com ID 1...");
-            reader = dbOperates.SelectById(1);
+            reader = repository.SelectById(1);
             if (reader.Read())
             {
                 Console.WriteLine($"Nome: {reader["Nome"]}, Idade: {reader["Idade"]}");
@@ -33,10 +35,10 @@ namespace SingletonCRUD
             reader.Close();
 
             Console.WriteLine("\nExcluindo registro com ID 2...");
-            dbOperates.Delete(2);
+            repository.Delete(1005);
 
             Console.WriteLine("\nExibindo todos os registros...");
-            reader = dbOperates.SelectAll();
+            reader = repository.SelectAll();
             while (reader.Read())
             {
                 Console.WriteLine($"Nome: {reader["Nome"]}, Idade: {reader["Idade"]}");
